backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
apscheduler==3.10.4
pydantic==2.6.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
httpx==0.26.0
scikit-learn==1.4.0
numpy==1.26.4
